+++ lib/src/misc/misc.dart
@@ -1,4 +1,5 @@
 export 'config_initializer.dart';
+export 'config_path.dart';
 export 'constants.dart';
 export 'cue_kinds.dart';
 export 'default_configs.dart';